syntax error fix

+++ Predictor/src/main/java/hardwar/branch/prediction/judged/PAp/PAp.java
@@ -30,7 +30,7 @@ public class PAp implements BranchPredictor {
 
         // Initializing the PAPHT with BranchInstructionSize as PHT Selector and 2^BHRSize row as each PHT entries
         // number and SCSize as block size
-        PAPHT = new PerAddressPredictionHistoryTable(branchInstructionSize, Math.pow(2, BHRSize), SCSize);
+        PAPHT = new PerAddressPredictionHistoryTable(branchInstructionSize, (int) Math.pow(2, BHRSize), SCSize);
 
         // Initialize the SC register
         SC = new SIPORegister("sc", SCSize, null);
